pr feedback, move ptr into sys which makes more sense
