+++ string/string.go
@@ -1,21 +1,16 @@
 package string
 
-import "strings"
+import (
+	"strings"
 
-// Ptr returns a pointer to the given value.
-func Ptr[T any](v T) *T {
-	return &v
-}
+	"github.com/agentuity/go-common/sys"
+)
 
 // StringPointer will set the pointer to nil if the string is not nil but an empty string
 func StringPointer(v string) *string {
 	if v != "" {
 		nv := strings.TrimSpace(v)
-		if nv == "" {
-			return nil
-		} else {
-			return &nv
-		}
+		return sys.Ptr(nv)
 	}
 	return nil
 }
@@ -27,7 +22,7 @@ func ClearEmptyStringPointer(v *string) *string {
 		if nv == "" {
 			return nil
 		} else {
-			return &nv
+			return sys.Ptr(nv)
 		}
 	}
 	return nil
